Fix: remove domain separator definition from constructor
EIP-2612 states that computing the domain separator in the constructor
creates a security vulnerability. Here's the excerpt:

"If the DOMAIN_SEPARATOR contains the chainId and is defined at contract
deployment instead of reconstructed for every signature, there is a risk
of possible replay attacks between chains in the event of a future chain
split."

So we're moving the construction of the domain separator to permit(),
where it belongs.